feat: Toggle balance between LTC and fiat currency
This commit introduces the ability for users to tap on their balance to toggle the display between Litecoin (LTC) and their selected fiat currency.

Key changes:
*   The main balance display is now clickable, allowing users to switch between the LTC and fiat currency values.
*   The `BalanceBentoGrid` no longer handles navigation clicks; this responsibility is now managed at a higher level.
*   A new `CurrentCurrencyStateFlow` has been introduced to observe changes in the user's selected currency.
*   The `PriceTickerGrid` has been updated to display the price ticker that corresponds to the currently selected currency, instead of showing a pager for all available pairs.
*   The `PriceTickerGrid` and `TransactionHistoryGrid` components and their associated state management have been moved into their own sub-packages for better organization.
