+++ app/src/main/java/com/brainwallet/ui/bento/BentoMainScreenState.kt
@@ -67,8 +67,4 @@ class BentoMainScreenState(
     fun toggleRail() {
         isRailOpen = !isRailOpen
     }
-
-    fun onBalanceClicked() {
-        ltcNavigation?.navigateToBuyLiteCoinScreen()
-    }
 }
