+++ modules/ltc/src/main/java/com/brainwallet/ltc/domain/flow/CurrentCurrencyStateFlow.kt
@@ -1,0 +1,9 @@
+package com.brainwallet.ltc.domain.flow
+
+import androidx.compose.runtime.Stable
+import kotlinx.coroutines.ExperimentalForInheritanceCoroutinesApi
+import kotlinx.coroutines.flow.StateFlow
+
+@OptIn(ExperimentalForInheritanceCoroutinesApi::class)
+@Stable
+interface CurrentCurrencyStateFlow : StateFlow<String>
